module rename, tidy up
